requirements.txt
alembic==1.10.2
argcomplete==3.0.4
autopep8==2.0.2
boto3==1.26.98
botocore==1.29.98
cachetools==5.3.0
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
durationpy==0.5
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
google-auth==2.17.0
hjson==3.1.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
placebo==0.9.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.7
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
troposphere==4.3.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zappa==0.56.1
zipp==3.15.0
